urdu_voice_ai/requirements.txt
Flask==2.3.3
faster-whisper==1.0.0
werkzeug==2.3.7
numpy==1.26.3